Extract `blur-field-accessibly` function
